참고한 자료의 연결부분을 내 데이터와 맞추어 오류발생하던 지점 해결

+++ src/main/java/com/example/jspboard2/domain/Member.java
@@ -7,7 +7,7 @@ import java.sql.Date;
 @Data
 public class Member {
 
-    private Integer id_member; // 멤버 ID
+    private Long id_member; // 멤버 ID
     private String user_member; // 멤버 작성되는 아이디정보(아이디)
     private String password_member; // 멤버 비밀번호
     private String name_member; // 멤버 이름
